requirements.txt
python-dotenv==1.0.0
filetype==1.2.0
loguru==0.7.0
pyyaml==6.0.1
wrapt==1.14.1